refactor(oracle): migrate ipsec scapy backends to ipsec plugin
